[QFEED-50] Refactor: Category 별도의 에그리거트로 분리 및 Question <-> Category 연관관계 해제
- Category는 User에서도 의존하는 엔티티라 분리하는게 좋다고 판단
- 다른 에그리거트와의 연관관계는 기본키로만 연관
