+++ module-domain/src/main/java/com/wsws/moduledomain/category/vo/CategoryName.java
@@ -1,4 +1,4 @@
-package com.wsws.moduledomain.feed.category.vo;
+package com.wsws.moduledomain.category.vo;
 
 public enum CategoryName {
     TRAVEL, DELICIOUS_RESTAURANT, MOVIE, MUSIC, READING, SPORTS
