+++ module-domain/src/main/java/com/wsws/moduledomain/feed/question/Question.java
@@ -1,10 +1,9 @@
 package com.wsws.moduledomain.feed.question;
 
-import com.wsws.moduledomain.feed.category.Category;
+import com.wsws.moduledomain.category.Category;
 import com.wsws.moduledomain.feed.question.vo.QuestionStatus;
 import jakarta.persistence.*;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
 
@@ -20,17 +19,15 @@ public class Question {
     private QuestionStatus questionStatus;
     private LocalDateTime createdAt;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "category_id")
-    private Category category;
+    private Long categoryId;
 
 
-    public static Question create(String content, QuestionStatus questionStatus, LocalDateTime createdAt, Category category) {
+    public static Question create(String content, QuestionStatus questionStatus, LocalDateTime createdAt, Long categoryId) {
         Question question = new Question();
         question.content = content;
         question.questionStatus = questionStatus;
         question.createdAt = createdAt;
-        question.category = category;
+        question.categoryId = categoryId;
         return question;
     }
